refactor(gateway): extract serve() into lib.rs
Move all server setup logic (adapter initialization, axum router,
background tasks, WebSocket handler) from main.rs into a
pub async fn serve(ServeConfig) in lib.rs.

The gateway binary is now a thin 15-line wrapper that calls
openab_gateway::serve(ServeConfig::default()).

This makes the gateway server embeddable and testable as a library.
